refactor: Change all native exceptions to custom exceptions.
Changed all IllegalStateException to IllegalUserAccessException in QuestionController to handle it by GlobalControllerExceptionHandler.
Changed all exceptions to UnauthorizedUserAccessException in UserController to handle it by GlobalControllerExceptionHandler.
Fixed User::isMatchedPassword to 'this.password' to compare password.
Changed GlobalControllerExceptionHandler::handleIllegalUserAccessException to remove Model, because it cannot get user list model when exception occurred.
Added Overloaded constructor in UnauthorizedUserAccessException to construct with a String message.