fix: Modified updateForm.html to render properly
It had the problem that is same with the problem form.html had, which is about type mismatch error.
It was caused by submit the value of 'value' property using DELETE method. So I moved them out of form tag.

Changed {{writer}} to {{writer.userId}} to render userId at view in show.html and updateForm.html.